Add and use log helpers
New log helper methods are used to add common information and context to log messages. All log messages are now prepended with broker information (host and port) as well as the client id. Also the word "MQTT" is now part of the common log output, making it obsolete for individual messages.

Co-Authored-By: Giovanni Giacobbi <thg2k@users.noreply.github.com>