Fix protocol implementation error when certain packets are bigger than 129 bytes
For some packets the remaining length was encoded using `chr()`, but if the packet exceeded 129 bytes (i.e. 127 remaining length) it would become incorrectly encoded. This changes fixes the problem by always using the proper encoder method `encodeMessageLength()`. Also drops the buffer length counter as it is faster to just check for `strlen()`.